+++ src/lib/Favorites.svelte
@@ -4,7 +4,6 @@
   import { goto } from '$app/navigation';
 
   import LL from '$src/i18n/i18n-svelte';
-  import { colors } from '$src/lib/colors';
   import { state } from '$src/stores';
 
   import Clapperboard from '~icons/lucide/clapperboard';
@@ -16,7 +15,6 @@
 
   import type { DisplayCredential } from '../../src-tauri/bindings/display-credential/DisplayCredential';
   import CredentialListItem from './credentials/CredentialListItem.svelte';
-  import { icons } from './icons';
 
   export let credentialType: 'all' | 'data' | 'badges' = 'all';
