+++ src/lib/events/History.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
   import type { Event, EventType } from '$lib/events';
+  import HistoryEntry from '$lib/events/HistoryEntry.svelte';
   import exampleEvents from '$src/lib/events/mock-data.json';
   import { state } from '$src/stores';
-
-  import HistoryEntry from './HistoryEntry.svelte';
 
   const events: Event[] = exampleEvents.map((e) => ({ ...e, type: e.type as EventType }));
   // const events: Event[] = [];
@@ -17,6 +16,7 @@
       lastConnected: 'n/a',
     },
     title: 'Initial connection',
+    timestamp: '2023-08-03T12:23:41.218Z',
     credentials: [],
   };
 
@@ -29,11 +29,12 @@
       lastConnected: 'n/a',
     },
     title: 'Data shared',
-    credentials: [$state.credentials[0]],
+    timestamp: '2023-08-03T12:23:42.749Z',
+    credentials: [$state.credentials[0], $state.credentials[1]],
   };
 </script>
 
-<div class="flex h-full flex-col">
+<div class="flex h-full flex-col items-end pr-4">
   {#if events.length === 0}
     <div class="flex h-full flex-col items-center justify-center">
       <p class="text-[14px]/[22px] font-medium text-slate-500 dark:text-slate-300">Coming soon ...</p>
