+++ src/routes/(app)/me/settings/profile/+page.svelte
@@ -3,11 +3,11 @@
 
   import { melt } from '@melt-ui/svelte';
 
+  import { dispatch } from '$lib/dispatcher';
   import LL from '$src/i18n/i18n-svelte';
   import Button from '$src/lib/components/atoms/Button.svelte';
   import BottomDrawer from '$src/lib/components/molecules/dialogs/BottomDrawer.svelte';
   import TopNavigation from '$src/lib/components/molecules/navigation/TopNavigation.svelte';
-  import { dispatch } from '$src/lib/dispatcher';
   import SettingsEntry from '$src/lib/settings/SettingsEntry.svelte';
 
   import Keyboard from '~icons/ph/keyboard-fill';
