+++ src/routes/welcome/password/completed/+page.svelte
@@ -6,9 +6,9 @@
 
   import '@lottiefiles/lottie-player';
 
+  import { dispatch } from '$lib/dispatcher';
   import LL from '$src/i18n/i18n-svelte';
   import Button from '$src/lib/components/atoms/Button.svelte';
-  import { dispatch } from '$src/lib/dispatcher';
   import { onboarding_state } from '$src/stores';
 
   import Shield from '~icons/ph/shield-fill';
