+++ src/routes/welcome/password/confirm/+page.svelte
@@ -4,10 +4,10 @@
 
   import { melt } from '@melt-ui/svelte';
 
+  import { dispatch } from '$lib/dispatcher';
   import LL from '$src/i18n/i18n-svelte';
   import Button from '$src/lib/components/atoms/Button.svelte';
   import TopNavigation from '$src/lib/components/molecules/navigation/TopNavigation.svelte';
-  import { dispatch } from '$src/lib/dispatcher';
   import { onboarding_state } from '$src/stores';
 
   import CheckCircle from '~icons/ph/check-circle-fill';
